requirements.txt
scipy==1.5.4
scikit-image==0.18.1
urllib3==1.25.8
opencv-contrib-python==4.2.0.32
tensorflow==2.3.1
opencv-python==4.2.0.32
Keras==2.4.3
flask==1.1.2
gunicorn
python-dotenv==0.15.0